Removed the preprocessor `REACHING RANGE` since it is no longer referenced inside of this library.
This library is now dependent on the **aInterfaceUtils** library and this library will not load if **aInterfaceUtils** is not found.

Removed all checks for **aInterfaceUtils** since this library is now dependent on it.

Removed some useless comments.

Removed the `toggleInventory` function. This should be defined by the developer if needed.

Removed the mouse events from `Diob/GroundItem`, these events should be defined by the developer for the best chance at customization.