Fix issue with stacked items
Fix a visual issue where if you picked up a full stack of items and added it to the inventory, if you had another non full stack in the inventory it would use the full stack you used to fill the non full stack, instead of just placing the full stack in a open slot. This caused extra calculations and looked weird visually because the stack did not go to the expected slot. (This was not broken) But this fix makes it work as expected and looks natural.